requirements.txt
google-generativeai==0.3.1
fastapi
pydantic
uvicorn
google-cloud-aiplatform
pyJWT
python-decouple
passlib
beanie
twilio